+++ src/main/kotlin/nhncommerce/project/redis/Constant/EventCoupon.kt
@@ -1,4 +1,4 @@
-package nhncommerce.project.radis.Constant
+package nhncommerce.project.redis.Constant
 
 enum class EventCoupon(val value : String) {
     COUPON("TimeCoupon")
