+++ src/main/kotlin/nhncommerce/project/redis/RedisRepository.kt
@@ -1,4 +1,4 @@
-package nhncommerce.project.radis
+package nhncommerce.project.redis
 
 import nhncommerce.project.user.domain.User
 import org.springframework.data.repository.CrudRepository
